+++ webApiReact/Models/UserReport.cs
@@ -53,13 +53,13 @@ namespace webApiReact.Models
         public string? adress { get; set; }
 
         [Column(TypeName = "bigint")]
-        public int? T_ON { get; set; }
+        public long? T_ON { get; set; }
 
         [Column(TypeName = "nvarchar(150)")]
         public string? F_I_O { get; set; }
 
         [Column(TypeName = "bigint")]
-        public int? K_NPO { get; set; }
+        public long? K_NPO { get; set; }
 
         public int? TIP2 { get; set; }
 
@@ -67,34 +67,34 @@ namespace webApiReact.Models
 
 
         [Column(TypeName = "bigint")]
-        public int? P1 { get; set; }
+        public long? P1 { get; set; }
 
         [Column(TypeName = "bigint")]
-        public int? P2 { get; set; }
+        public long? P2 { get; set; }
 
         [Column(TypeName = "bigint")]
-        public int? p3 { get; set; }
+        public long? p3 { get; set; }
 
         [Column(TypeName = "bigint")]
-        public int? p4 { get; set; }
+        public long? p4 { get; set; }
 
         [Column(TypeName = "bigint")]
-        public int? p5 { get; set; }
+        public long? p5 { get; set; }
 
         [Column(TypeName = "bigint")]
-        public int? p6 { get; set; }
+        public long? p6 { get; set; }
 
         [Column(TypeName = "bigint")]
-        public int? p7 { get; set; }
+        public long? p7 { get; set; }
 
         [Column(TypeName = "bigint")]
-        public int? p8 { get; set; }
+        public long? p8 { get; set; }
 
         [Column(TypeName = "bigint")]
-        public int? p9 { get; set; }
+        public long? p9 { get; set; }
 
         [Column(TypeName = "bigint")]
-        public int? p10 { get; set; }
+        public long? p10 { get; set; }
 
         [Column(TypeName = "bit")]
         public bool? E002 { get; set; }
